move util out of journeys folder
