+++ x-pack/performance/journeys_e2e/web_logs_dashboard_esql_long_running.ts
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-import { createDashboardJourney } from './utils/dashboard_journey';
+import { createDashboardJourney } from '../utils/dashboard_journey';
 
 export const journey = createDashboardJourney({
   esArchives: ['x-pack/performance/es_archives/sample_data_logs_many_fields'],
