+++ x-pack/performance/utils/dashboard_journey.ts
